added psql profile and module tags

+++ models/staging/staging_sales.sql
@@ -9,7 +9,7 @@ WITH raw_sales AS (
         CAST(sale_id AS INTEGER) AS sale_id,
         CAST(product_id AS INTEGER) AS product_id,
         CAST(customer_id AS INTEGER) AS customer_id,
-        CAST(amount AS DOUBLE) AS amount,
+        CAST(amount AS DOUBLE PRECISION) AS amount,
         CAST(date AS DATE) AS date
     FROM {{ source('raw', 'sales') }}
 )
